requirements.txt
flask==3.0.0
youtube-transcript-api==0.6.2
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
